requirements.txt
pyyaml>=6.0.1
sqlalchemy>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0 